Add periodic charges to budget
The suite needs to periodically charge the amount
specified on a periodic movement to the designed budget.

To do this, a periodic task scheduler such as celery can be used,
by defining all the tasks required to register the charges when
the periodic movement is created, and then actually modifying
the budget's available amount when the period defined ends.